Support loop/resolution-change decoding.
Previously, when stop_streaming function was called, the decoder
instance was destroyed immediately.
Fixed do not destroy the instance immediately.
Changed to store the write/read pointer of input/output buffer when
the stop_streaming function was called. The write/read pointer are
used in the start sequence if host want to decode continuously.
Added the VPU interface to control display flag and inform the read
pointer.

Signed-off-by: Nas Chung <nas.chung@chipsnmedia.com>