media: chips-media: wave5: Add hrtimer based polling support
Add support for starting a polling timer in case interrupt
is not available. This helps keep the VPU functional in
SoC's such as AM62A, where the hardware interrupt
hookup may not be present due to an SoC errata [1].

The timer is shared across all instances of encoder and
decoder and is started when first instance of encoder
or decoder is started.

hrtimer callback is called with 5ms polling interval
while any of the encoder/decoder instances are running to check
the interrupt status as being done in irq handler.

Based on above interrupt status, use a worker thread to
iterate over the interrupt status for each instance
and send completion event as being done in irq thread function.

Parse for irq number before v4l2 device registration and
if not available initialize hrtimer and worker thread.

Protect hrtimer access and instance list with device specific
mutex locks to avoid race conditions while different instances
of encoder and decoder are started together.

Add module param to change polling interval for debug purpose.

[1] Errata ID: i2373 : "VPU's interrupt is not routed to A53"

Co-developed-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Devarsh Thakkar <devarsht@ti.com>